requirements.txt
Flask==3.0.0
python-docx==1.1.0
qrcode==7.4.2
Pillow==10.0.0
openai==1.3.0